Split kernel operations into its own class
Each NotebookClient can now produce multiple kernels, thus
matching the notebook REST API. These can be used with
async context managers, so we always clean up the kernel
when done.

We move OperationError out to break cyclic imports
