+++ binderbot/__init__.py
@@ -7,3 +7,7 @@ __version__ = '0.1.0'
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
+
+
+class OperationError(Exception):
+    pass
